+++ mod2/helm-generator/helmchartgenerator-core/src/test/java/org/onap/dcaegen2/platform/helmchartgenerator/ComponentSpecValidatorTest.java
@@ -37,7 +37,7 @@ public class ComponentSpecValidatorTest {
 
     @BeforeEach
     void setUp() {
-        validator = new ComponentSpecValidatorImpl();
+        validator = new ComponentSpecValidatorImpl(new Utils());
     }
 
     @Test